+++ mobile/src/screens/MyProfile/index.tsx
@@ -21,12 +21,12 @@ import api from '@services/api';
 import { createBlankClassSchedule } from '@utils/factories';
 import { parseFetchedToParsedClassSchedule } from '@utils/mappers';
 
-import { ClassScheduleForm } from '@components/ClassScheduleForm';
+import { ClassScheduleForm } from '@components/form/ClassScheduleForm';
 import { FormContainer } from '@components/form/FormContainer';
+import { FormFieldset } from '@components/form/FormFieldset';
 import { OuterLabelInput } from '@components/form/OuterLabelInput';
 import { Select } from '@components/form/Select';
 import { TextArea } from '@components/form/TextArea';
-import { FormFieldset } from '@components/FormFieldset';
 import { ScreenHeader } from '@components/ScreenHeader';
 import { UserAvatar } from '@components/UserAvatar';
